Añadir clases de monstruos y reorganizar código
Se han añadido las clases `CowboyMonster`, `Dracula` y `Outlaw` en `Monster.hpp` y sus implementaciones en `Monster.cpp`, mejorando la gestión de los comportamientos de los monstruos.

Se han agregado declaraciones de amistad en `PrairieKing.hpp` para permitir el acceso a miembros privados.

Se han eliminado las implementaciones de las clases de monstruos de `PrairieKing.cpp`, lo que mejora la organización del código.

Además, se han ajustado las funciones de actualización de música y se han añadido nuevas inclusiones de encabezados para facilitar el uso de las nuevas clases.
