+++ include/gameplay/PrairieKing.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "AssetManager.hpp"
+#include "gameplay/Monster.hpp"
 #include "raylib.h"
 #include <vector>
 #include <unordered_map>
@@ -52,12 +53,13 @@ namespace std {
 
 class PrairieKing {
 public:
+    friend class CowboyMonster;
+    friend class Dracula;
+    friend class Outlaw;
+
     // Forward declarations
-    class CowboyMonster;
     class CowboyBullet;
     class CowboyPowerup;
-    class Dracula;
-    class Outlaw;
     class JOTPKProgress;
 
     // Enum declarations
@@ -235,87 +237,6 @@ public:
                                bool flip, float depth, Color color);
         bool Update(float deltaTime);
         void Draw(const Texture2D& texture);
-    };
-
-    class CowboyMonster {
-    public:
-        static const int MONSTER_ANIMATION_DELAY = 500;
-
-        int health;
-        int type;
-        int speed;
-        float movementAnimationTimer;
-        Rectangle position;
-        int movementDirection;
-        bool movedLastTurn;
-        bool oppositeMotionGuy;
-        bool invisible;
-        bool special;
-        bool uninterested;
-        bool flyer;
-        Color tint;
-        Color flashColor;
-        float flashColorTimer;
-        int ticksSinceLastMovement;
-        Vector2 acceleration;
-        Vector2 targetPosition;
-
-        CowboyMonster(int which, int health, int speed, Vector2 position);
-        CowboyMonster(int which, Vector2 position);
-        virtual void Draw(const Texture2D& texture, Vector2 topLeftScreenCoordinate);
-        virtual bool TakeDamage(int damage);
-        virtual int GetLootDrop();
-        virtual bool Move(Vector2 playerPosition, float deltaTime);
-        void SpikeyEndBehavior(int extraInfo);
-    };
-
-    class Dracula : public CowboyMonster {
-    public:
-        static const int GLOATING_PHASE = -1;
-        static const int WALK_RANDOMLY_AND_SHOOT_PHASE = 0;
-        static const int SPREAD_SHOT_PHASE = 1;
-        static const int SUMMON_DEMON_PHASE = 2;
-        static const int SUMMON_MUMMY_PHASE = 3;
-
-        int phase;
-        int phaseInternalTimer;
-        int phaseInternalCounter;
-        int shootTimer;
-        int fullHealth;
-        Vector2 homePosition;
-
-        Dracula();
-        void Draw(const Texture2D& texture, Vector2 topLeftScreenCoordinate) override;
-        int GetLootDrop() override;
-        bool TakeDamage(int damage) override;
-        bool Move(Vector2 playerPosition, float deltaTime) override;
-        void FireSpread(Vector2 origin, double offsetAngle);
-        void SummonEnemies(Vector2 origin, int which);
-    };
-
-    class Outlaw : public CowboyMonster {
-    public:
-        static const int TALKING_PHASE = -1;
-        static const int HIDING_PHASE = 0;
-        static const int DART_OUT_AND_SHOOT_PHASE = 1;
-        static const int RUN_AND_GUN_PHASE = 2;
-        static const int RUN_GUN_AND_PANT_PHASE = 3;
-        static const int SHOOT_AT_PLAYER_PHASE = 4;
-
-        int phase;
-        int phaseCountdown;
-        int shootTimer;
-        int phaseInternalTimer;
-        int phaseInternalCounter;
-        bool dartLeft;
-        int fullHealth;
-        Vector2 homePosition;
-
-        Outlaw(Vector2 position, int health);
-        void Draw(const Texture2D& texture, Vector2 topLeftScreenCoordinate) override;
-        bool Move(Vector2 playerPosition, float deltaTime) override;
-        int GetLootDrop() override;
-        bool TakeDamage(int damage) override;
     };
 
     PrairieKing(AssetManager& assets);
